Move EfExecutor and EfSubExecutor to Vdf
We initially noticed something needed to change here because testVdfSpeculation
needs to use EfExecutor, in order to use VdfParallelExecutorEngine. The
downstream dependency causes problems in some builds. That pointed out that
VdfExecutor is core functionality that it makes sense for Vdf to provide, and
VdfSubExecutor is a dependency that comes along for the ride.

We need to follow up on this by adding Vdf unit test coverage for the moved
classes, but there are no relevant tests in ef that can be moved along with
these files. So we're proceeding with moving these definitions in order to fix
the problem with the testVdfSpeculation dependencies.

Fixes #3871

(Internal change: 2401163)
