+++ pxr/exec/ef/pageCacheSubExecutor.h
@@ -12,10 +12,10 @@
 #include "pxr/pxr.h"
 
 #include "pxr/exec/ef/pageCacheBasedExecutor.h"
-#include "pxr/exec/ef/subExecutor.h"
 
 #include "pxr/exec/vdf/executorFactory.h"
 #include "pxr/exec/vdf/speculationExecutor.h"
+#include "pxr/exec/vdf/subExecutor.h"
 
 PXR_NAMESPACE_OPEN_SCOPE
 
@@ -50,7 +50,7 @@ class EfPageCacheSubExecutor :
     // Executor factory.
     typedef
         VdfExecutorFactory<
-            EfSubExecutor<EngineType, DataManagerType>,
+            VdfSubExecutor<EngineType, DataManagerType>,
             VdfSpeculationExecutor<SpeculationEngineType, DataManagerType>>
         _Factory;
