+++ pxr/exec/vdf/executor.h
@@ -4,8 +4,8 @@
 // Licensed under the terms set forth in the LICENSE.txt file available at
 // https://openusd.org/license.
 //
-#ifndef PXR_EXEC_EF_SUB_EXECUTOR_H
-#define PXR_EXEC_EF_SUB_EXECUTOR_H
+#ifndef PXR_EXEC_VDF_EXECUTOR_H
+#define PXR_EXEC_VDF_EXECUTOR_H
 
 ///\file
 
@@ -13,11 +13,12 @@
 
 #include "pxr/base/tf/mallocTag.h"
 #include "pxr/base/trace/trace.h"
-#include "pxr/exec/vdf/dataManagerBasedSubExecutor.h"
+#include "pxr/exec/vdf/dataManagerBasedExecutor.h"
 #include "pxr/exec/vdf/executorFactory.h"
 #include "pxr/exec/vdf/executorInterface.h"
 #include "pxr/exec/vdf/request.h"
 #include "pxr/exec/vdf/speculationExecutor.h"
+#include "pxr/exec/vdf/subExecutor.h"
 
 PXR_NAMESPACE_OPEN_SCOPE
 
@@ -25,55 +26,45 @@ class VdfExecutorErrorLogger;
 class VdfSchedule;
 
 ///////////////////////////////////////////////////////////////////////////////
-//
-// \class EfSubExecutor.
-//
-// \brief Executed a VdfNetwork to compute a requested set of value, and uses
-//        cached output values from a parent executor, if unavailable in the
-//        local data manager.
-//
+///
+/// \class VdfExecutor
+///
+/// \brief Executes a VdfNetwork to compute a requested set of values.
+///
+///
 template <
     template <typename> class EngineType,
     typename DataManagerType>
-class EfSubExecutor : 
-    public VdfDataManagerBasedSubExecutor<DataManagerType, VdfExecutorInterface>
+class VdfExecutor : 
+    public VdfDataManagerBasedExecutor<DataManagerType, VdfExecutorInterface>
 {
-    // Base type definition
+    // Base class type.
     typedef
-        VdfDataManagerBasedSubExecutor<DataManagerType, VdfExecutorInterface>
+        VdfDataManagerBasedExecutor<DataManagerType, VdfExecutorInterface>
         Base;
 
-    // The speculation executor engine alias declaration, to be bound as a 
+    // The speculation executor engine alias declaration, to be bound as a
     // template template parameter.
     template <typename T>
     using SpeculationEngineType =
         typename EngineType<T>::SpeculationExecutorEngine;
 
     // Executor factory.
-    typedef
+    typedef 
         VdfExecutorFactory<
-            EfSubExecutor<EngineType, DataManagerType>,
+            VdfSubExecutor<EngineType, DataManagerType>,
             VdfSpeculationExecutor<SpeculationEngineType, DataManagerType>>
         _Factory;
 
 public:
 
-    /// Default constructor
+    /// Default constructor.
     ///
-    EfSubExecutor() :
-        _engine(*this, &this->_dataManager)
-    { }
+    VdfExecutor() : _engine(*this, &this->_dataManager) {}
 
-    /// Construct with a parent executor
+    /// Destructor.
     ///
-    explicit EfSubExecutor(const VdfExecutorInterface *parentExecutor) :
-        Base(parentExecutor),
-        _engine(*this, &this->_dataManager)
-    { }
-
-    /// Destructor
-    ///
-    virtual ~EfSubExecutor() {}
+    virtual ~VdfExecutor() {}
 
     /// Factory construction.
     ///
@@ -81,20 +72,20 @@ public:
         return _factory;
     }
 
-protected:
+private:
 
-    /// Run this executor with the given \p schedule and \p request.
-    ///
+    // Run this executor with the given \p schedule and \p request.
+    //
     virtual void _Run(
         const VdfSchedule &schedule,
         const VdfRequest &computeRequest,
         VdfExecutorErrorLogger *errorLogger) override;
 
-private:
-    
-    /// Clears the data in the data manager.
-    ///
+    // Causes the DataManager to clear its temporary execution buffers.
+    //
     virtual void _ClearData() override;
+
+private:
 
     // The factory shared amongst executors of this type.
     //
@@ -103,45 +94,37 @@ private:
     // This is the engine that will do most of our hard work for us.
     //
     EngineType<DataManagerType> _engine;
-
 };
 
 ///////////////////////////////////////////////////////////////////////////////
 
 template <template <typename> class EngineType, typename DataManagerType>
-const typename EfSubExecutor<EngineType, DataManagerType>::_Factory
-    EfSubExecutor<EngineType, DataManagerType>::_factory;
+const typename VdfExecutor<EngineType, DataManagerType>::_Factory
+    VdfExecutor<EngineType, DataManagerType>::_factory;
 
 template <template <typename> class EngineType, typename DataManagerType>
 void
-EfSubExecutor<EngineType, DataManagerType>::_Run(
+VdfExecutor<EngineType, DataManagerType>::_Run(
     const VdfSchedule &schedule,
     const VdfRequest &computeRequest,
     VdfExecutorErrorLogger *errorLogger)
-{
-    // If we have an empty request, bail out.
+{    
+    // If we have an empty request, bail.
     if (computeRequest.IsEmpty()) {
         return;
     }
 
     TRACE_FUNCTION();
-    TfAutoMallocTag2 tag("Ef", "EfSubExecutor::Run");
+    TfAutoMallocTag2 tag("Ef", "VdfExecutor::Run");
 
     _engine.RunSchedule(schedule, computeRequest, errorLogger);
 }
 
 template <template <typename> class EngineType, typename DataManagerType>
 void
-EfSubExecutor<EngineType, DataManagerType>::_ClearData()
+VdfExecutor<EngineType, DataManagerType>::_ClearData()
 {
-    TRACE_FUNCTION();
-
-    // If the data manager is empty, don't even attempt to clear it.
-    if (!Base::_dataManager.IsEmpty()) {
-        Base::_dataManager.Clear();
-    }
-
-    Base::InvalidateTopologicalState();
+    Base::_dataManager.Clear();
 }
 
 PXR_NAMESPACE_CLOSE_SCOPE
